requirements.txt
Flask==2.3.3
pylti==0.7.1
requests==2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0